+++ src/main/java/innexgo/CourseRowMapper.java
@@ -9,9 +9,9 @@ public class CourseRowMapper implements RowMapper<Course> {
   @Override
   public Course mapRow(ResultSet row, int rowNum) throws SQLException {
     Course course = new Course();
-    course.id = row.getInt("id");
-    course.teacherId = row.getInt("teacher_id");
-    course.locationId = row.getInt("location_id");
+    course.id = row.getLong("id");
+    course.teacherId = row.getLong("teacher_id");
+    course.locationId = row.getLong("location_id");
     course.period = row.getInt("period");
     course.subject = row.getString("subject");
     course.year = row.getInt("year");
