+++ src/main/java/innexgo/LocationRowMapper.java
@@ -9,7 +9,7 @@ public class LocationRowMapper implements RowMapper<Location> {
   @Override
   public Location mapRow(ResultSet row, int rowNum) throws SQLException {
     Location loc = new Location();
-    loc.id = row.getInt("id");
+    loc.id = row.getLong("id");
     loc.name = row.getString("name");
     loc.tags = row.getString("tags");
     return loc;
