+++ src/main/java/innexgo/PeriodRowMapper.java
@@ -9,7 +9,7 @@ public class PeriodRowMapper implements RowMapper<Period> {
   @Override
   public Period mapRow(ResultSet row, int rowNum) throws SQLException {
     Period period = new Period();
-    period.id = row.getInt("id");
+    period.id = row.getLong("id");
     period.initialTime = row.getLong("initial_time");
     period.startTime = row.getLong("start_time");
     period.endTime = row.getLong("end_time");
