+++ src/main/java/innexgo/ScheduleService.java
@@ -13,27 +13,27 @@ public class ScheduleService {
 
   @Autowired private JdbcTemplate jdbcTemplate;
 
-  public Schedule getById(int id) {
+  public Schedule getById(long id) {
     String sql = "SELECT id, student_id, course_id FROM schedule WHERE id=?";
     RowMapper<Schedule> rowMapper = new ScheduleRowMapper();
     List<Schedule> schedules = jdbcTemplate.query(sql, rowMapper, id);
     return schedules.size() == 0 ? null : schedules.get(0);
   }
 
-  public Schedule getScheduleByStudentIdCourseId(int studentId, int courseId) {
+  public Schedule getScheduleByStudentIdCourseId(long studentId, long courseId) {
     String sql = "SELECT id, student_id, course_id FROM schedule WHERE student_id=? AND course_id=?";
     RowMapper<Schedule> rowMapper = new ScheduleRowMapper();
     List<Schedule> schedules = jdbcTemplate.query(sql, rowMapper, studentId, courseId);
     return schedules.size() == 0 ? null : schedules.get(0);
   }
 
-  public boolean existsById(int id) {
+  public boolean existsById(long id) {
     String sql = "SELECT count(*) FROM schedule WHERE id=?";
     int count = jdbcTemplate.queryForObject(sql, Integer.class, id);
     return count != 0;
   }
 
-  public boolean existsByStudentIdCourseId(int studentId, int courseId) {
+  public boolean existsByStudentIdCourseId(long studentId, long courseId) {
     String sql = "SELECT count(*) FROM schedule WHERE student_id=? AND course_id=?";
     int count = jdbcTemplate.queryForObject(sql, Integer.class, studentId, courseId);
     return count != 0;
@@ -55,7 +55,7 @@ public class ScheduleService {
 
       // Fetch schedule id
       sql = "SELECT id FROM schedule WHERE student_id=? AND course_id=?";
-      int id = jdbcTemplate.queryForObject(sql, Integer.class, schedule.studentId, schedule.courseId);
+      long id = jdbcTemplate.queryForObject(sql, Long.class, schedule.studentId, schedule.courseId);
       schedule.id = id;
       return schedule;
     } else {
@@ -68,7 +68,7 @@ public class ScheduleService {
     jdbcTemplate.update(sql, schedule.id, schedule.studentId, schedule.courseId, schedule.id);
   }
 
-  public Schedule deleteById(int id) {
+  public Schedule deleteById(long id) {
     Schedule schedule = getById(id);
     String sql = "DELETE FROM schedule WHERE id=?";
     jdbcTemplate.update(sql, id);
@@ -76,11 +76,11 @@ public class ScheduleService {
   }
 
   public List<Schedule> query(
-      Integer scheduleId,
-      Integer studentId,
-      Integer courseId,
-      Integer teacherId,
-      Integer locationId,
+      Long scheduleId,
+      Long studentId,
+      Long courseId,
+      Long teacherId,
+      Long locationId,
       Integer period) {
     String sql =
         "SELECT s.id, s.student_id, s.course_id FROM schedule s"
